modified issue where playlist and playlist track was causing circular dependency, need to implement playlisttracksreference to get the playlisttrack
