+++ backend/src/types/playlistTrack.ts
@@ -1,14 +1,16 @@
-// import { z } from 'zod';
-// import { Track } from './track.ts';
-// import { User } from './user.ts';
+import { z } from 'zod';
+import { Track } from './track.ts';
+import { User } from './user.ts';
 
-// export const PlaylistTrack = z.object({
-//     id: z.number(),
-//     addedAt: z.date(),
-//     addedBy: User,
-//     track: Track,
-//     // Removed since we dont need access to the playlist
-//     // playlist: Playlist
-// });
+// Required as a standalone function
+export const PlaylistTrack = z.object({
+    id: z.number(),
+    addedAt: z.date(),
+    addedBy: User,
+    track: Track,
+    // Removed since we dont need access to the playlist
+    // Also this will lead to circular dependency so dont use it
+    // playlist: Playlist
+});
 
-// export type PlaylistTrackZ = z.input<typeof PlaylistTrack>;
+export type PlaylistTrackZ = z.input<typeof PlaylistTrack>;
